Refactor contributor activity ingestion and tests
