+++ src/hiero_analytics/data_sources/rate_limit.py
@@ -11,9 +11,11 @@ from __future__ import annotations
 import logging
 from collections.abc import Mapping
 from dataclasses import dataclass
-from datetime import datetime, timezone
+from datetime import UTC, datetime
 from enum import Enum, auto
 from typing import Any
+
+import requests
 
 logger = logging.getLogger(__name__)
 
@@ -23,6 +25,7 @@ JSON = dict[str, Any]
 # --------------------------------------------------------
 # NORMALIZED SNAPSHOT
 # --------------------------------------------------------
+
 
 @dataclass(frozen=True)
 class RateLimitSnapshot:
@@ -37,9 +40,9 @@ class RateLimitSnapshot:
     only has to be written once.
     """
 
-    remaining: int | None = None    # requests / points left
-    limit: int | None = None        # total budget per window
-    cost: int | None = None         # GraphQL only: cost of this query
+    remaining: int | None = None  # requests / points left
+    limit: int | None = None  # total budget per window
+    cost: int | None = None  # GraphQL only: cost of this query
     reset_at: datetime | None = None  # aware UTC datetime when budget resets
 
     @classmethod
@@ -64,11 +67,7 @@ class RateLimitSnapshot:
         except (TypeError, ValueError):
             return None
 
-        reset_at = (
-            datetime.fromtimestamp(reset_epoch, tz=timezone.utc)
-            if reset_epoch is not None
-            else None
-        )
+        reset_at = datetime.fromtimestamp(reset_epoch, tz=UTC) if reset_epoch is not None else None
 
         return cls(remaining=remaining, limit=limit, reset_at=reset_at)
 
@@ -99,12 +98,13 @@ class RateLimitSnapshot:
         """Seconds until the budget window resets (0 if unknown or already passed)."""
         if self.reset_at is None:
             return 0
-        return max(0, int((self.reset_at - datetime.now(timezone.utc)).total_seconds()))
+        return max(0, int((self.reset_at - datetime.now(UTC)).total_seconds()))
 
 
 # --------------------------------------------------------
 # DECISION
 # --------------------------------------------------------
+
 
 class Action(Enum):
     """What the GitHubClient loop should do after reading rate-limit signals."""
@@ -134,7 +134,8 @@ class Action(Enum):
 
 @dataclass(frozen=True)
 class RateLimitDecision:
-    """ A decision returned by the policy after inspecting a snapshot or response.
+    """A decision returned by the policy after inspecting a snapshot or response.
+
     The GitHubClient loop will apply the decision by sleeping and/or retrying
     as needed.
     """
@@ -156,11 +157,68 @@ _GRAPHQL_LOW_BUDGET_SLEEP = 5
 _GRAPHQL_ERROR_FALLBACK_SLEEP = 300
 # GraphQL RATE_LIMIT error: never sleep less than this even if reset_at is soon
 _GRAPHQL_ERROR_MIN_SLEEP = 60
+_SECONDARY_RATE_LIMIT_FALLBACK_SLEEP_SECONDS = 60
+_SECONDARY_RATE_LIMIT_MARKERS = (
+    "secondary rate limit",
+    "abuse detection",
+    "please wait a few minutes",
+    "temporarily blocked",
+)
+
+
+def response_error_message(response: requests.Response) -> str:
+    """Extract a short API error message from a failed HTTP response."""
+    try:
+        payload = response.json()
+    except ValueError:
+        text = response.text.strip()
+        return text[:300]
+
+    if isinstance(payload, dict):
+        message = payload.get("message")
+        if isinstance(message, str) and message.strip():
+            return message.strip()
+
+        errors = payload.get("errors")
+        if isinstance(errors, list) and errors:
+            return str(errors[0])
+
+    return ""
+
+
+def retry_after_seconds(response: requests.Response) -> int:
+    """Parse GitHub's Retry-After header when present."""
+    raw_retry_after = response.headers.get("Retry-After")
+    if raw_retry_after is None:
+        return 0
+
+    try:
+        return max(0, int(raw_retry_after))
+    except ValueError:
+        return 0
+
+
+def secondary_rate_limit_sleep_seconds(response: requests.Response) -> int | None:
+    """Return a retry delay for retryable 403 responses."""
+    if response.status_code != 403:
+        return None
+
+    retry_after = retry_after_seconds(response)
+    message = response_error_message(response).lower()
+
+    if retry_after > 0:
+        return retry_after
+
+    if any(marker in message for marker in _SECONDARY_RATE_LIMIT_MARKERS):
+        return _SECONDARY_RATE_LIMIT_FALLBACK_SLEEP_SECONDS
+
+    return None
 
 
 # --------------------------------------------------------
 # POLICY
 # --------------------------------------------------------
+
 
 class RateLimitPolicy:
     """
@@ -234,6 +292,7 @@ class RateLimitPolicy:
     ) -> RateLimitDecision:
         """
         Proactive back-off when GraphQL budget is critically low.
+
         Called after every successful GraphQL response.
         """
         if snapshot.remaining is None:
