+++ src/hiero_analytics/run_maintainer_pipeline_org.py
@@ -3,9 +3,8 @@ Run contributor responsibility analytics for a GitHub org.
 
 Produces:
 - Raw contributor activity log
-- Actor stage journeys
-- Yearly general user -> triage -> maintainer pipeline
-- Repository-level highest observed stage distribution
+- Org-wide actor stage journeys
+- Yearly contributor responsibility pipeline
 """
 
 from __future__ import annotations
@@ -13,13 +12,14 @@ from __future__ import annotations
 import os
 from collections.abc import Callable, Sequence
 
+import pandas as pd
+
+from hiero_analytics.analysis.dataframe_utils import contributor_activity_to_dataframe
 from hiero_analytics.analysis.maintainer_pipeline import (
     STAGE_LABELS,
     STAGE_ORDER,
-    activity_records_to_dataframe,
     build_cumulative_stage_timeline,
     build_maintainer_pipeline,
-    build_repo_stage_distribution,
     build_stage_activity_timeline,
     build_stage_entry_timeline,
     summarize_actor_stage_journeys,
@@ -27,7 +27,7 @@ from hiero_analytics.analysis.maintainer_pipeline import (
 from hiero_analytics.config.charts import RESPONSIBILITY_COLORS
 from hiero_analytics.config.paths import ORG, ensure_org_dirs
 from hiero_analytics.data_sources.github_client import GitHubClient
-from hiero_analytics.data_sources.github_ingest import (
+from hiero_analytics.data_sources.github_contributor_activity import (
     DEFAULT_CONTRIBUTOR_ACTIVITY_LOOKBACK_DAYS,
     fetch_org_contributor_activity_graphql,
 )
@@ -37,10 +37,7 @@ from hiero_analytics.plotting.bars import plot_stacked_bar
 from hiero_analytics.plotting.lines import plot_multiline
 
 STACK_LABELS = [STAGE_LABELS[stage] for stage in STAGE_ORDER]
-RESPONSIBILITY_STAGE_COLORS = {
-    STAGE_LABELS[stage]: RESPONSIBILITY_COLORS[STAGE_LABELS[stage]]
-    for stage in STAGE_ORDER
-}
+RESPONSIBILITY_STAGE_COLORS = {STAGE_LABELS[stage]: RESPONSIBILITY_COLORS[STAGE_LABELS[stage]] for stage in STAGE_ORDER}
 DEFAULT_CONTRIBUTOR_ACTIVITY_CACHE_TTL_SECONDS = 86_400
 
 
@@ -55,7 +52,7 @@ def resolve_activity_max_workers() -> int:
 
 
 def resolve_activity_lookback_days() -> int | None:
-    """Resolve the repo-relative contributor-activity lookback from the environment."""
+    """Resolve the contributor-activity lookback from the environment."""
     raw_value = os.getenv(
         "GITHUB_CONTRIBUTOR_ACTIVITY_LOOKBACK_DAYS",
         str(DEFAULT_CONTRIBUTOR_ACTIVITY_LOOKBACK_DAYS),
@@ -112,35 +109,39 @@ def print_maintainer_runtime_settings(
     if activity_lookback_days is None:
         output_fn("Using contributor activity lookback: full history")
     else:
-        output_fn(
-            "Using contributor activity lookback: "
-            f"{activity_lookback_days} days from each repo's latest issue or PR update"
-        )
+        output_fn(f"Using contributor activity lookback: {activity_lookback_days} days")
     output_fn(f"Using contributor activity cache TTL: {activity_cache_ttl_seconds}s")
     output_fn(f"Using pause between repo fetches: {repo_pause_seconds:g}s")
     if selected_repos:
         output_fn(f"Restricting contributor activity fetch to {len(selected_repos)} repo(s)")
 
 
+def _stage_timeline_to_long(stage_timeline: pd.DataFrame, *, period_col: str) -> pd.DataFrame:
+    """Convert a wide stage timeline into the shared long plotting shape."""
+    return stage_timeline.melt(
+        id_vars=[period_col],
+        value_vars=STAGE_ORDER,
+        var_name="stage",
+        value_name="count",
+    ).assign(stage=lambda dataframe: dataframe["stage"].map(STAGE_LABELS))
+
+
 def save_maintainer_pipeline_outputs(activities: list[ContributorActivityRecord]) -> None:
     """Build and save maintainer pipeline tables and charts from fetched activity."""
     org_data_dir, org_charts_dir = ensure_org_dirs(ORG)
 
-    activity_df = activity_records_to_dataframe(activities)
-    repo_stage_journeys = summarize_actor_stage_journeys(activities, by_repo=True)
-    org_stage_journeys = summarize_actor_stage_journeys(activities, by_repo=False)
-    yearly_pipeline = build_maintainer_pipeline(org_stage_journeys)
-    monthly_pipeline = build_stage_entry_timeline(org_stage_journeys, frequency="month")
+    activity_df = contributor_activity_to_dataframe(activities)
+    stage_journeys = summarize_actor_stage_journeys(activities)
+    yearly_pipeline = build_maintainer_pipeline(stage_journeys)
+    monthly_pipeline = build_stage_entry_timeline(stage_journeys, frequency="month")
     cumulative_monthly_pipeline = build_cumulative_stage_timeline(
         monthly_pipeline,
         period_col="month",
     )
     monthly_stage_activity = build_stage_activity_timeline(activities, frequency="month")
-    repo_stage_distribution = build_repo_stage_distribution(repo_stage_journeys)
 
     save_dataframe(activity_df, org_data_dir / "maintainer_activity_log.csv")
-    save_dataframe(repo_stage_journeys, org_data_dir / "maintainer_stage_journeys_by_repo.csv")
-    save_dataframe(org_stage_journeys, org_data_dir / "maintainer_stage_journeys_org.csv")
+    save_dataframe(stage_journeys, org_data_dir / "maintainer_stage_journeys_org.csv")
     save_dataframe(yearly_pipeline, org_data_dir / "maintainer_pipeline_yearly.csv")
     save_dataframe(monthly_pipeline, org_data_dir / "maintainer_pipeline_monthly.csv")
     save_dataframe(
@@ -150,10 +151,6 @@ def save_maintainer_pipeline_outputs(activities: list[ContributorActivityRecord]
     save_dataframe(
         monthly_stage_activity,
         org_data_dir / "maintainer_stage_activity_monthly.csv",
-    )
-    save_dataframe(
-        repo_stage_distribution,
-        org_data_dir / "maintainer_stage_distribution_by_repo.csv",
     )
 
     print("Saved maintainer pipeline tables")
@@ -169,30 +166,9 @@ def save_maintainer_pipeline_outputs(activities: list[ContributorActivityRecord]
             output_path=org_charts_dir / "maintainer_pipeline_yearly.png",
         )
 
-    if not repo_stage_distribution.empty:
-        plot_stacked_bar(
-            repo_stage_distribution,
-            x_col="repo",
-            stack_cols=STAGE_ORDER,
-            labels=STACK_LABELS,
-            colors=RESPONSIBILITY_COLORS,
-            title="Highest Observed Contributor Stage by Repository",
-            output_path=org_charts_dir / "maintainer_stage_distribution_by_repo.png",
-            rotate_x=45,
-        )
-
     if not monthly_pipeline.empty:
-        monthly_pipeline_long = (
-            monthly_pipeline.melt(
-                id_vars=["month"],
-                value_vars=STAGE_ORDER,
-                var_name="stage",
-                value_name="count",
-            )
-            .assign(stage=lambda df: df["stage"].map(STAGE_LABELS))
-        )
         plot_multiline(
-            monthly_pipeline_long,
+            _stage_timeline_to_long(monthly_pipeline, period_col="month"),
             x_col="month",
             y_col="count",
             group_col="stage",
@@ -203,17 +179,8 @@ def save_maintainer_pipeline_outputs(activities: list[ContributorActivityRecord]
         )
 
     if not cumulative_monthly_pipeline.empty:
-        cumulative_monthly_pipeline_long = (
-            cumulative_monthly_pipeline.melt(
-                id_vars=["month"],
-                value_vars=STAGE_ORDER,
-                var_name="stage",
-                value_name="count",
-            )
-            .assign(stage=lambda df: df["stage"].map(STAGE_LABELS))
-        )
         plot_multiline(
-            cumulative_monthly_pipeline_long,
+            _stage_timeline_to_long(cumulative_monthly_pipeline, period_col="month"),
             x_col="month",
             y_col="count",
             group_col="stage",
@@ -224,22 +191,13 @@ def save_maintainer_pipeline_outputs(activities: list[ContributorActivityRecord]
         )
 
     if not monthly_stage_activity.empty:
-        monthly_stage_activity_long = (
-            monthly_stage_activity.melt(
-                id_vars=["month"],
-                value_vars=STAGE_ORDER,
-                var_name="stage",
-                value_name="count",
-            )
-            .assign(stage=lambda df: df["stage"].map(STAGE_LABELS))
-        )
         plot_multiline(
-            monthly_stage_activity_long,
+            _stage_timeline_to_long(monthly_stage_activity, period_col="month"),
             x_col="month",
             y_col="count",
             group_col="stage",
             colors=RESPONSIBILITY_STAGE_COLORS,
-            title="Active Contributors by Responsibility Signal per Month",
+            title="Active Contributors by Attained Responsibility Stage per Month",
             output_path=org_charts_dir / "maintainer_stage_activity_monthly_line.png",
             rotate_x=45,
         )
@@ -276,7 +234,6 @@ def main() -> None:
 
     print(f"Fetched {len(activities)} contributor activity records")
     save_maintainer_pipeline_outputs(activities)
-
     print("Maintainer pipeline analytics complete")
