+++ tests/test_run_analytics.py
@@ -5,32 +5,34 @@ from __future__ import annotations
 import hiero_analytics.run_analytics as run_analytics
 
 
-def _build_test_pipelines(call_log: list[str]) -> tuple[run_analytics.PipelineSpec, ...]:
+def _build_test_pipelines(call_log: list[str]) -> tuple[dict[str, object], ...]:
     """Create a small fake pipeline registry for launcher tests."""
     return (
-        run_analytics.PipelineSpec(
-            key="onboarding",
-            label="Onboarding",
-            description="Onboarding analytics.",
-            runner=lambda: call_log.append("onboarding"),
-            aliases=("gfi",),
-        ),
-        run_analytics.PipelineSpec(
-            key="difficulty",
-            label="Difficulty",
-            description="Difficulty analytics.",
-            runner=lambda: call_log.append("difficulty"),
-        ),
-        run_analytics.PipelineSpec(
-            key="maintainer",
-            label="Maintainer",
-            description="Maintainer analytics.",
-            runner=lambda: call_log.append("maintainer"),
-        ),
+        {
+            "key": "onboarding",
+            "label": "Onboarding",
+            "description": "Onboarding analytics.",
+            "runner": lambda: call_log.append("onboarding"),
+            "aliases": ("gfi",),
+        },
+        {
+            "key": "difficulty",
+            "label": "Difficulty",
+            "description": "Difficulty analytics.",
+            "runner": lambda: call_log.append("difficulty"),
+            "aliases": (),
+        },
+        {
+            "key": "maintainer",
+            "label": "Maintainer",
+            "description": "Maintainer analytics.",
+            "runner": lambda: call_log.append("maintainer"),
+            "aliases": (),
+        },
     )
 
 
-def test_parse_selection_supports_numbers_and_names(monkeypatch):
+def test_parse_selection_supports_numbers_and_names(monkeypatch) -> None:
     """Prompt parsing should accept menu numbers and pipeline names."""
     call_log: list[str] = []
     monkeypatch.setattr(run_analytics, "PIPELINES", _build_test_pipelines(call_log))
@@ -40,7 +42,7 @@ def test_parse_selection_supports_numbers_and_names(monkeypatch):
     assert selection == ["onboarding", "maintainer"]
 
 
-def test_prompt_for_selection_retries_after_invalid_input(monkeypatch):
+def test_prompt_for_selection_retries_after_invalid_input(monkeypatch) -> None:
     """Interactive prompts should loop until the user enters a valid choice."""
     call_log: list[str] = []
     monkeypatch.setattr(run_analytics, "PIPELINES", _build_test_pipelines(call_log))
@@ -57,7 +59,7 @@ def test_prompt_for_selection_retries_after_invalid_input(monkeypatch):
     assert any(message.startswith("Invalid selection:") for message in output)
 
 
-def test_main_runs_selected_pipelines_once(monkeypatch, capsys):
+def test_main_runs_selected_pipelines_once(monkeypatch, capsys) -> None:
     """Named CLI selection should run each chosen pipeline once in order."""
     call_log: list[str] = []
     monkeypatch.setattr(run_analytics, "PIPELINES", _build_test_pipelines(call_log))
@@ -71,7 +73,7 @@ def test_main_runs_selected_pipelines_once(monkeypatch, capsys):
     assert "[1/2] Running Difficulty analytics" in captured.out
 
 
-def test_main_all_runs_every_pipeline(monkeypatch):
+def test_main_all_runs_every_pipeline(monkeypatch) -> None:
     """The all flag should execute the full registry in menu order."""
     call_log: list[str] = []
     monkeypatch.setattr(run_analytics, "PIPELINES", _build_test_pipelines(call_log))
